Reorder "should" section and remove plugin method
Reordered the 'should' section so it is _now_ in order of how the bot
would typically start. I've also removed the plugin load/unload method
of using file write detection. The reason for this change is because
having plugins load/unload based on file writing is potentially risky,
and might cause problems with the bot itself.